# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MADAD community help platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface User {
  id: string;
  email: string;
  username: string;
  displayName: string;
  role: string;
  emailVerified?: boolean;
}

export interface CurrentUserResponse {
  user: User | null;
}

export interface AuthResponse {
  user: User;
  verificationRequired?: boolean;
}

export interface RegisterInput {
  email: string;
  /**
     * @minLength 3
     * @maxLength 24
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 120
     */
  password: string;
  /** @maxLength 80 */
  displayName?: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type PostAuthor = {
  username: string;
};

export type _PostCount = {
  comments: number;
  helpOffers: number;
};

export interface Post {
  id: string;
  title: string;
  content: string;
  category: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  city?: string | null;
  status: string;
  createdAt: string;
  author: PostAuthor;
  _count: _PostCount;
}

export interface PostInput {
  /**
     * @minLength 5
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 1200
     */
  content: string;
  category: string;
  /** @maxLength 80 */
  country?: string;
  /** @maxLength 80 */
  city?: string;
}

export interface PostListResponse {
  posts: Post[];
}

export interface PostResponse {
  post: Post;
}

export interface HelpOfferInput {
  /** @maxLength 500 */
  message?: string;
}

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 800
     */
  content: string;
}

export interface ReportInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  reason: string;
  /** @maxLength 800 */
  details?: string;
}

/**
 * Error response
 */
export type ErrorResponseResponse = Error;

export type ListPostsParams = {
category?: string;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};