# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MADAD community help platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Session and account operations
  - name: posts
    description: Community request operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current session user
      responses:
        "200":
          description: Current user or an empty session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new MADAD account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Sign in to a MADAD account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed-in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Sign out of the current session
      responses:
        "200":
          description: Session ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List active community requests
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 30
      responses:
        "200":
          description: Active community requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      operationId: createPost
      tags: [posts]
      summary: Publish a community request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Request published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /posts/{postId}/help:
    post:
      operationId: offerHelp
      tags: [posts]
      summary: Offer help on a request
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpOfferInput"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /posts/{postId}/comments:
    post:
      operationId: addComment
      tags: [posts]
      summary: Add a helpful comment
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /posts/{postId}/report:
    post:
      operationId: reportPost
      tags: [posts]
      summary: Report a community request
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
components:
  parameters:
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        username: { type: string }
        displayName: { type: string }
        role: { type: string }
        emailVerified: { type: boolean }
      required: [id, email, username, displayName, role]
    CurrentUserResponse:
      type: object
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
      required: [user]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        verificationRequired:
          type: boolean
      required: [user]
    RegisterInput:
      type: object
      properties:
        email: { type: string, format: email }
        username: { type: string, minLength: 3, maxLength: 24 }
        password: { type: string, minLength: 8, maxLength: 120 }
        displayName: { type: string, maxLength: 80 }
      required: [email, username, password]
    LoginInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [email, password]
    Post:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        content: { type: string }
        category: { type: string }
        country: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        author:
          type: object
          properties:
            username: { type: string }
          required: [username]
        _count:
          type: object
          properties:
            comments: { type: integer }
            helpOffers: { type: integer }
          required: [comments, helpOffers]
      required: [id, title, content, category, status, createdAt, author, _count]
    PostInput:
      type: object
      properties:
        title: { type: string, minLength: 5, maxLength: 120 }
        content: { type: string, minLength: 10, maxLength: 1200 }
        category: { type: string }
        country: { type: string, maxLength: 80 }
        city: { type: string, maxLength: 80 }
      required: [title, content, category]
    PostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
      required: [posts]
    PostResponse:
      type: object
      properties:
        post:
          $ref: "#/components/schemas/Post"
      required: [post]
    HelpOfferInput:
      type: object
      properties:
        message: { type: string, maxLength: 500 }
    CommentInput:
      type: object
      properties:
        content: { type: string, minLength: 1, maxLength: 800 }
      required: [content]
    ReportInput:
      type: object
      properties:
        reason: { type: string, minLength: 1, maxLength: 160 }
        details: { type: string, maxLength: 800 }
      required: [reason]