# lib/api-zod/src/generated/types/currentUserResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MADAD community help platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface CurrentUserResponse {
  user: User | null;
}