# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MADAD community help platform API
 * OpenAPI spec version: 0.1.0
 */
import type { _PostCount } from './_postCount';
import type { PostAuthor } from './postAuthor';

export interface Post {
  id: string;
  title: string;
  content: string;
  category: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  city?: string | null;
  status: string;
  createdAt: Date;
  author: PostAuthor;
  _count: _PostCount;
}