# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MADAD community help platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /**
     * @minLength 5
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 1200
     */
  content: string;
  category: string;
  /** @maxLength 80 */
  country?: string;
  /** @maxLength 80 */
  city?: string;
}