# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MADAD community help platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  email: string;
  /**
     * @minLength 3
     * @maxLength 24
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 120
     */
  password: string;
  /** @maxLength 80 */
  displayName?: string;
}