# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MADAD community help platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  reason: string;
  /** @maxLength 800 */
  details?: string;
}